requirements.txt
pytesseract
python-Levenshtein
